script: Move common source set operations into `srcset.rs`
